Reworked J4M_efficiency
Changed the calculation and the docstring to use the embedded functions for Si, Al, and Kapton. Turns out the measured efficiency was not measured, it was calculated from other soruces. Therefore, this function will now rely on tabulated values rather than the curve on the sheet. This way, the function can now take in a theta arugment to correctly account for the detector efficiency at scattering angles larger than 0.