Added azimithalBinning to utils, changed geometry_calibration
azimuthalBinning is modeled off of Silke's azimuthalBinning.py in smalldatatools, but I refactored it and re-worked the normalization to happen before the masking. Removed a redundant ELEMENT_NUMBERS macro. Removed phi option from run_geometry_calibration(). Turns our that the Pplane variable in smalldatatools' azimuthalBinning.py was not an azimuthal offset to correct for a detector rotation relative to the polarization of incoming pulses. It is actually an incoherent mixing factor from 0-1 of vertically to horizontally polarized light. Makes more sense when dealing with synchrotrons when this is a more prominent thing. However for the LCLS-I (not LCLS-II as APPLE-II undulators can deliver arbitrary polarizations) the polarization to a very good approximation is either linear or vertrtical, not a mix. Therefore this parameter should only be set to 1 or 0 for us in the producer. Most of the time 0. To be fair, the polarization corrections that we get from the 5D fit were on the order of less than 1 percent, so I think it is fine to leave it out, especially if we cannot account for it in the producer. Further modification of azimuthalBinning.py from smalldatatools would be needed. If the time comes, that can be implemented. For now, I leave it out.