Updated combineRuns, added get_data_paths
combineRuns now supports different file locations for different runs. The folders kwarg can be a string, character array, or list of either. Old functionality is retained. If you have a list of runs, you can dictate one folder. You can also dictate a list of folders for multiple runs. The length of the runs and the list of folders must be equal if the length of the list of folders is greater than 1. get_data_paths is a function which takes in a run number or array of run numbers, and returns a list of folders for each run number, where the locations are specified by a .yaml file. The path to the .yaml file defaults to , though it can be specified as a kwarg. Also updated docstring for combineRuns.